🔨 Refactor(#142): 튜토리얼에서 소개로 도메인 수정
